packages/app_btc/pyproject.toml
[tool.poetry]
name = "cypherock-sdk-app-btc"
version = "0.1.0"
description = "Bitcoin app for Cypherock SDK"
authors = ["Temiloluwa Yusuf <ytemiloluwa@gmail.com>"]
readme = "README.md"
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = ">=3.11"
python-bitcointx = "^1.1.5"
btclib = "^2023.7.12"
httpx = "^0.24.0"

[tool.poetry.group.dev.dependencies]
ruff = "^0.5.0"
black = "^24.4.2"
mypy = "^1.10.0"
pytest = "^8.2.2"
pytest-asyncio = "^1.1.0"

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 88
target-version = ['py311']

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.pytest.ini_options]
pythonpath = "."
filterwarnings = [
    "ignore::pytest.PytestRemovedIn9Warning",
    "ignore::DeprecationWarning",
    "ignore::RuntimeWarning",
    "ignore:.*Task exception was never retrieved.*:RuntimeWarning",
    "ignore:.*coroutine.*was never awaited.*:RuntimeWarning",
]
